Move some things to tree utils
